Update README.md
This marks a milestone with SM.  Hatch integration with CLS was a long desired feature for me, and took awhile.  I'm very happy with the results, and with the updates to CLS, highlighting issues are now a thing of the past.

I've added a lot of polish to the mod with this release, and it has been very stable in test (a lot of test :D).