Fix multi-crew transfers (crew were being lost), but KIS inventory is only transferred for some of the crew.
Also some refactoring of the crew movement code to get rid of duplicate and dead code, and makes things slightly more understandable.

Still need to figure out why multi-crew transfers break KIS Inventory. Maybe we need to schedule one move per update. A way to achieve this would be to build a list of crew members to move, and then move one per update. To get rid of the constant polling we should ideally do this in a coroutine. In general that would probably be a better approach throughout SM.